+++ R/utils.R
@@ -1,4 +1,4 @@
-#' @import kableExtra magrittr knitr utils
+#' @import magrittr knitr utils
 
 
 
@@ -10,6 +10,28 @@
 if(!exists("dynare") || !is.environment(dynare)) dynare<<-new.env()
 }
 
+.onAttach <- function(libname, pkgname) {
+  packageStartupMessage("Thank you for using DynareR!
+
+          To acknowledge our work, please cite the package:
+
+                        PLAIN TEXT:
+
+  Sagiru Mati (2019). DynareR: Bringing the Power of Dynare to R, R
+  Markdown, and Quarto. CRAN. https://CRAN.R-project.org/package=DynareR
+
+              BIBTEX:
+
+
+  @Article{Mati2019,
+    title = {DynareR: Bringing the Power of {Dynare} to {R}, {R Markdown}, and {Quarto}},
+    author = {Sagiru Mati},
+    year = {2019},
+    journal = {CRAN},
+    url = {https://CRAN.R-project.org/package=DynareR},
+  }")
+}
+
 
 
 # dir_create
@@ -18,7 +40,7 @@ dir_create=function(x) if(!dir.exists(x)) dir.create(x,recursive = T)
 
 #  globalVariables
 
-globalVariables(c("."))
+globalVariables(c("." ))
 
 # run_model
 
@@ -45,9 +67,9 @@ run_model <- function(model,import_log=F) {
 
   octaveFile<-basename(tempfile(model, '.',".m"))   # m is file extension of octave/matlab
 
-  if(!exists("addPath")) set_dynare_version()
+  if(!exists("matlabPath")) set_dynare_version()
 
-  writeLines(c(addPath,paste0('cd ',modelDir),sprintf("dynare %s",dynarePath)), octaveFile)
+  writeLines(c(matlabPath,paste0('cd ',modelDir),sprintf("dynare %s",dynarePath)), octaveFile)
 
 
   on.exit(unlink(octaveFile),add = T)
@@ -59,11 +81,12 @@ run_model <- function(model,import_log=F) {
 # system_exec
 
 system_exec=function(){
-  if(!exists("engine_path")) set_octave_path()
-  engine_path=eval(expression(engine_path),envir = parent.frame())
+  if(!exists("octavePath")) set_octave_path()
+  octavePath=eval(expression(octavePath),envir = parent.frame())
   octaveFile=eval(expression(octaveFile),envir = parent.frame()) # Dynamic scoping
-  system2(set_octave_path(engine_path),paste("--eval",shQuote(paste("run",octaveFile))))
-}
+  if(octavePath=="")  stop('Please provide the correct path to the Octave executable compatible with the Dynare Version')
+  system2(set_octave_path(octavePath),paste("--eval",shQuote(paste("run",octaveFile))))
+  }